Fix #101738
