+++ src/vs/platform/keybinding/common/keybindingResolver.ts
@@ -3,9 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { isNonEmptyArray } from 'vs/base/common/arrays';
-import { isIMenuItem, MenuId, MenuRegistry } from 'vs/platform/actions/common/actions';
-import { CommandsRegistry, ICommandHandlerDescription } from 'vs/platform/commands/common/commands';
 import { IContext, ContextKeyExpression, ContextKeyExprType } from 'vs/platform/contextkey/common/contextkey';
 import { ResolvedKeybindingItem } from 'vs/platform/keybinding/common/resolvedKeybindingItem';
 
@@ -339,44 +336,6 @@ export class KeybindingResolver {
 			return true;
 		}
 		return rules.evaluate(context);
-	}
-
-	public static getAllUnboundCommands(boundCommands: Map<string, boolean>): string[] {
-		const unboundCommands: string[] = [];
-		const seenMap: Map<string, boolean> = new Map<string, boolean>();
-		const addCommand = (id: string, includeCommandWithArgs: boolean) => {
-			if (seenMap.has(id)) {
-				return;
-			}
-			seenMap.set(id, true);
-			if (id[0] === '_' || id.indexOf('vscode.') === 0) { // private command
-				return;
-			}
-			if (boundCommands.get(id) === true) {
-				return;
-			}
-			if (!includeCommandWithArgs) {
-				const command = CommandsRegistry.getCommand(id);
-				if (command && typeof command.description === 'object'
-					&& isNonEmptyArray((<ICommandHandlerDescription>command.description).args)) { // command with args
-					return;
-				}
-			}
-			unboundCommands.push(id);
-		};
-
-		// Add all commands from Command Palette
-		for (const menuItem of MenuRegistry.getMenuItems(MenuId.CommandPalette)) {
-			if (isIMenuItem(menuItem)) {
-				addCommand(menuItem.command.id, true);
-			}
-		}
-
-		for (const id of CommandsRegistry.getCommands().keys()) {
-			addCommand(id, false);
-		}
-
-		return unboundCommands;
 	}
 }
